DROTH-3514 Move old samuutus code from LaneService to LaneUpdater
